Move PermissionsSection into components
